修改ModbusRegReadDefinitions结构体的字段类型，更新ModbusWidget以支持文件的打开和保存功能，添加注册别名和格式的获取与设置方法

+++ ModbusRegReadDefinitions.h
@@ -8,8 +8,8 @@ struct ModbusRegReadDefinitions
     quint8 id;
     quint8 function;
     quint16 reg_addr;
-    quint32 quantity;
-    quint32 scan_rate;
+    quint16 quantity;
+    qint32 scan_rate;
     QByteArray packet;
 };
